TD9-NEW-03: split commands.rs into per-domain submodules
commands.rs had grown to 2425 LOC as a flat collection of 27
independent `impl ConsoleCommand` structs plus shared formatting
helpers — past the 2000-LOC ceiling tracked in TD9-NEW-03 (part of
the #1323 complexity inventory).

Decompose into a `commands/` module along the issue's proposed
per-command-group axis. Each command is self-contained; the only
coupling is the formatting/lookup helpers and the external-type
imports, both re-exported from a `shared` prelude that every
submodule pulls in via a single `use super::shared::*` glob (glob
imports are exempt from the partial-unused-import warning, so the
zero-warning gate holds without per-file curated import lists):

  commands/mod.rs        build_command_registry() + module wiring (64)
  commands/shared.rs     import prelude + 9 cross-command helpers (179)
  commands/world_info.rs help/stats/entities/systems/sys.accesses/... (368)
  commands/assets.rs     tex.*/mesh.*/skin.* diagnostics (618)
  commands/view.rs       prid/cam.*/near/pick (527)
  commands/scene.rs      light.*/door.teleport/mat.*/ragdoll/script (704)

Largest file drops 2425 -> 704 LOC. Behavior-preserving: the public
surface (build_command_registry) is unchanged, all 27 commands stay
registered in the same order, and the existing commands_tests.rs runs
unmodified (pulled in via #[path] from mod.rs). 459 byroredux tests
pass; full workspace suite green; zero warnings.

Part of #1323.
